Move JavaFX Models functions into Models
Move JavaFX Models functions into Models so that there isn't an empty ModelsKt class generated
for Java interop.
